Move evaluate.py tests to their own file
